==> Analysis Scripts/04_analysis_and_results.Rmd <==
---
title: "Results"
output:
  html_document:
    df_print: paged
  pdf_document: default
  word_document: default
---

```{r setup, include=FALSE}
if (!require("ggplot2")) {
  install.packages("ggplot2")
  require("ggplot2")
}
if (!require("dplyr")) {
  install.packages("dplyr")
  require("dplyr")
}
if (!require("tidyr")) {
  install.packages("tidyr")
  require("tidyr")
}
if (!require("knitr")) {
  install.packages("knitr")
  require("knitr")
}
if (!require("patchwork")) {
  install.packages("patchwork")
  require("patchwork")
}
if (!require("mgcv")) {
  install.packages("mgcv")
  require("mgcv")
}
if (!require("lme4")) {
  install.packages("lme4")
  require("lme4")
}
if (!require("lmerTest")) {
  install.packages("lmerTest")
  require("lmerTest")
}
if (!require("gamm4")) {
  install.packages("gamm4")
  require("gamm4")
}
if (!require("ggbeeswarm")) {
  install.packages("ggbeeswarm")
  require("ggbeeswarm")
}
if (!require("emmeans")) {
  install.packages("emmeans")
  require("emmeans")
}
if (!require("MuMIn")) {
  install.packages("MuMIn")
  require("MuMIn")
}
if (!require("interactions")) {
  install.packages("interactions")
  require("interactions")
}
```

```{r echo=FALSE, warning=FALSE}
clean_data <- read.csv("/tsd/p274/data/durable/projects/p027-cr_bm/Clean Data All Cohorts/Wide Format/Analysed data.csv")
crbm_data <- read.csv("/tsd/p274/data/durable/projects/p027-cr_bm/crbm_data.csv")

crbm_data <- crbm_data %>% 
  select(id, CR, BM)

clean_data <- merge(clean_data, crbm_data, by = c("id"), all = TRUE) 

p_value_df <- read.csv("/tsd/p274/data/durable/projects/p027-cr_bm/p_value.csv")

# Change the p-values to none scientific format
p_value_df$adjusted_p_value <- as.numeric(format(p_value_df$adjusted_p_value, scientific = FALSE))
```

```{r echo=FALSE}
participants <- clean_data %>% nrow()
female <- (clean_data %>% filter(clean_data$sex == "female") %>% nrow())
female_percentage <- (female / participants) * 100
mean_age <- mean(clean_data$age_baseline, na.rm = TRUE)
sd_age <- sd(clean_data$age_baseline, na.rm = TRUE)

# Create a new variable to represent average follow-up time
clean_data <- clean_data %>%
  mutate(FU_time = rowMeans(select(., mem_time, hc_time, fc_time), na.rm = TRUE)) %>% 
  filter(!is.na(CR))

mean_fu <- mean(clean_data$FU_time, na.rm = TRUE)
sd_fu <- sd(clean_data$FU_time, na.rm = TRUE)
```

```{r echo=FALSE}
# Helper function to format p-values
format_p_value <- function(p) {
  if (p < 0.0001) {
    return("<0.0001") # Written like this because of HTML formatting
  } else {
    return(round(p, 4))
  }
}

# Helper function to perform ANOVA and extract required statistics
get_aov_stats <- function(formula, data, vars_priority) {
  model <- aov(formula, data = data)
  summary_model <- summary(model)
  
  # Access the ANOVA table
  anova_table <- summary_model[[1]]
  
  # Clean up row names to remove whitespace
  rownames(anova_table) <- trimws(rownames(anova_table))
  
  # Check each variable in priority order and break on the first match
  for (var in vars_priority) {
    if (var %in% rownames(anova_table)) {
      p_value <- anova_table[var, "Pr(>F)"]
      return(list(p_value = p_value))
    }
  }
  
  stop("None of the priority variables are in the model.")
}

# Set a list of priority variables
vars <- c("cohort")

# Call the function
mmse_model <- get_aov_stats(mmse ~ cohort, data = clean_data, vars_priority = vars)

# Age baseline
age_model <- get_aov_stats(age_baseline ~ cohort, data = clean_data, vars)
age_model_fu <- get_aov_stats(age_followup ~ cohort, data = clean_data, vars)

# Sex
chisq_test <- chisq.test(clean_data$sex, clean_data$cohort)
chisq_test$p_value <- chisq_test$p.value

# Education
edu_model <- get_aov_stats(edu ~ cohort, data = clean_data, vars)

# FU time
fu_time_model <- get_aov_stats(FU_time ~ cohort, data = clean_data, vars)

# FU observations
fu_obs_model <- get_aov_stats(FU_obs ~ cohort, data = clean_data, vars)
```

We included `r participants` participants: `r round(female_percentage, 2)`% female with a mean baseline age of `r round(mean_age, 2)` years (SD = `r round(sd_age, 2)`) and a mean follow-up time of `r round(mean_fu, 2)` years (SD = `r round(sd_fu, 2)`). The descriptive statistics are summarized by cohort in Table 3. At least two cohorts varied significantly on baseline age (_p_ = `r format_p_value(age_model$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "age_model", "adjusted_p_value"])`), follow-up age (_p_ = `r format_p_value(age_model_fu$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "age_model_fu", "adjusted_p_value"])`), sex (_p_ = `r format_p_value(chisq_test$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "chisq_test", "adjusted_p_value"])`), education (_p_ = `r format_p_value(edu_model$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "edu_model", "adjusted_p_value"])`), MMSE (_p_ = `r format_p_value(mmse_model$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "mmse_model", "adjusted_p_value"])`), follow-up time (_p_ = `r format_p_value(fu_time_model$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "fu_time_model", "adjusted_p_value"])`), and follow-up observations (_p_ = `r format_p_value(fu_obs_model$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "fu_obs_model", "adjusted_p_value"])`). We examined the memory composite harmonization across cohorts in Supplementary Figure 1. We found that the normalized memory composite scores across cohorts have broadly comparable distributions, with means near zero and standard deviations close to one.

```{r echo=FALSE}
# Create a new variable to represent average number of follow-ups
clean_data <- clean_data %>%
  mutate(FU_obs = rowMeans(select(., em_obs, hc_obs, fc_obs), na.rm = TRUE))

clean_data <- clean_data %>%
  mutate(Study = dplyr::recode(cohort,
    "Betula" = "Umea Betula",
    "Cobra" = "Umea COBRA",
    "Oslo" = "Oslo LCBC",
    "WAHA" = "Barcelona WAHA"
  ))

summary_stats <- clean_data %>%
  # Create a custom order for the table
  mutate(Study = factor(Study, levels = c("Barcelona WAHA", "Oslo LCBC", "Umea Betula", "Umea COBRA"))) %>%
  arrange(Study)

summary_stats <- summary_stats %>%
  dplyr::group_by(Study) %>%
  dplyr::summarise(
    `**Participants**` = n_distinct(id),
    `**Baseline age [mean (range)]**` = paste0(
      round(mean(age_baseline, na.rm = T), digits = 1),
      " (", paste0(round(range(age_baseline, na.rm = T)),
        collapse = "-"
      ),
      ")"
    ),
    `**Follow-up age [mean (range)]**` = paste0(
      round(mean(age_followup, na.rm = T), digits = 1),
      " (", paste0(round(range(age_followup, na.rm = T)),
        collapse = "-"
      ),
      ")"
    ),
    `**Sex [%F/%M]**` = paste0(
      c(
        round(mean(sex == "female", na.rm = T) * 100),
        round(mean(sex == "male", na.rm = T) * 100)
      ),
      collapse = "/"
    ),
    `**Education [mean (range)]**` = paste0(
      round(mean(edu, na.rm = T), digits = 1),
      " (", paste0(round(range(edu, na.rm = T)),
        collapse = "-"
      ),
      ")"
    ),
    `**MMSE [mean (range)]**` = paste0(
      round(mean(mmse, na.rm = T), digits = 1),
      " (", paste0(round(range(mmse, na.rm = T)),
        collapse = "-"
      ),
      ")"
    ),
    `**Follow-up time [mean (range)]**` = paste0(
      round(mean(FU_time, na.rm = T), digits = 1),
      " (", paste0(round(range(FU_time, na.rm = T)),
        collapse = "-"
      ),
      ")"
    ),
    `**Follow-up observations [mean (range)]**` = paste0(
      round(mean(FU_obs, na.rm = TRUE), digits = 1),
      " (",
      "2",
      "-",
      round(max(FU_obs, na.rm = TRUE)),
      ")"
    )
  )

# Create a data frame for the total line
total_line <- tibble::tibble(
  `Study` = "Total",
  `**Participants**` = n_distinct(clean_data$id),
  `**Baseline age [mean (range)]**` = paste0(
    round(mean(clean_data$age_baseline, na.rm = T), digits = 1),
    " (", paste0(round(range(clean_data$age_baseline, na.rm = T)),
      collapse = "-"
    ),
    ")"
  ),
  `**Follow-up age [mean (range)]**` = paste0(
    round(mean(clean_data$age_followup, na.rm = T), digits = 1),
    " (", paste0(round(range(clean_data$age_followup, na.rm = T)),
      collapse = "-"
    ),
    ")"
  ),
  `**Sex [%F/%M]**` = paste0(
    c(
      round(mean(clean_data$sex == "female", na.rm = T) * 100),
      round(mean(clean_data$sex == "male", na.rm = T) * 100)
    ),
    collapse = "/"
  ),
  `**Education [mean (range)]**` = paste0(
    round(mean(clean_data$edu, na.rm = T), digits = 1),
    " (", paste0(round(range(clean_data$edu, na.rm = T)),
      collapse = "-"
    ),
    ")"
  ),
  `**MMSE [mean (range)]**` = paste0(
    round(mean(clean_data$mmse, na.rm = T), digits = 1),
    " (", paste0(round(range(clean_data$mmse, na.rm = T)),
      collapse = "-"
    ),
    ")"
  ),
  `**Follow-up time [mean (range)]**` = paste0(
    round(mean(clean_data$FU_time, na.rm = T), digits = 1),
    " (", paste0(round(range(clean_data$FU_time, na.rm = T)),
      collapse = "-"
    ),
    ")"
  ),
  `**Follow-up observations [mean (range)]**` = paste0(
    round(mean(clean_data$FU_obs, na.rm = TRUE), digits = 1),
    " (",
    "2",
    "-",
    round(max(clean_data$FU_obs, na.rm = TRUE)),
    ")"
  )
)

# Bind the total line data frame to the summary statistics table
summary_stats <- bind_rows(summary_stats, total_line)
```

__Table 3__: _Demographic characteristics_

```{r tab3, echo=FALSE}
knitr::kable(summary_stats, align = "c")
```

<small>_Note._ Abbreviations: MMSE Mini Mental State Examination, F female, M male, CI Confidence Interval.</small>

We found that as age increased both memory and hippocampus volume decreased, though we also observed inter-individual differences in these relationships (Supplementary Figure 1). Both the LME model and GAMM fit to this data showed evidence of significant relationships between memory and hippocampal volume regressed on age; however, there was evidence of a non-linear relationship in both models (Supplementary Table 1). We include further details in the supplementary materials.

__**Episodic Memory and Hippocampus Change:**__ 

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Helper function to perform regression and extract required statistics
get_regression_stats <- function(formula, data, vars_priority) {
  model <- lm(formula, data = data)
  summary_model <- summary(model)
  coefficients <- summary_model$coefficients

  # Initialize the variable index
  variable_index <- NULL

  # Check for each variable in priority order and break on the first match
  for (var in vars_priority) {
    if (var %in% rownames(coefficients)) {
      variable_index <- which(rownames(coefficients) == var)
      break
    }
  }

  # Check if a relevant variable was found
  if (is.null(variable_index)) {
    stop("None of the priority variables are in the model.")
  }

  coef <- coefficients[variable_index, 1]
  ci <- confint(model)[variable_index, ]
  p_value <- coefficients[variable_index, 4]
  t_stat <- coefficients[variable_index, 3]
  adj_r2 <- summary_model$adj.r.squared
  r2 <- summary_model$r.squared

  return(list(coef = coef, ci = ci, p_value = p_value, t_stat = t_stat, adj_r2 = adj_r2, r2 = r2))
}

# Set a list of priority variables
hc_vars <- c(
  "scale(hc_slopes)",
  "hc_slopes",
  "scale(adj_HIP_total_baseline)",
  "scale(adj_HIP_total_followup)"
)

edu_var <- c(
  "scale(edu)"
)

# Running the regression model
emhp_merged <- get_regression_stats(scale(memory_slopes) ~ scale(hc_slopes) + study + scale(age_baseline) + scale(FU_time) + sex + scale(edu), data = clean_data, hc_vars)
emhp_merged_unscaled <- get_regression_stats(memory_slopes ~ hc_slopes + study + age_baseline + FU_time + sex + edu, data = clean_data, hc_vars)

# Cross-sectional analysis at baseline
model_baseline <- get_regression_stats(scale(memory_baseline) ~ scale(adj_HIP_total_baseline) + study + scale(age_baseline) + scale(edu) + sex, data = clean_data, hc_vars)

# Cross-sectional analysis at follow-up
# In order to make sure that the follow-up hc and memory values are from the same timepoint (eg because someone has more memory obs than hc obs), filter for participants to only those with values from the same timepoint
clean_data_followup <- clean_data %>%
  rowwise() %>%
  # Find which memory.x column equals memory_followup etc.
  mutate(
    memory_idx = match(memory_followup, c_across(starts_with("memory."))),   
    hc_idx     = match(adj_HIP_total_followup,     c_across(starts_with("adj_HIP_total."))), 
    age_idx    = match(age_followup,    c_across(starts_with("age.")))       
  ) %>%
  ungroup() %>%
  # Only keep rows where the matched memory.x, hc.x, and age.x column indices are the same
  filter(memory_idx == hc_idx, memory_idx == age_idx)

model_followup <- get_regression_stats(scale(memory_followup) ~ scale(adj_HIP_total_followup) + study + scale(age_followup) + scale(edu), data = clean_data_followup, hc_vars)
```

Hippocampus and memory annual change were significantly correlated (β = `r round(emhp_merged$coef, 4)`, _p_ = `r format_p_value(emhp_merged$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "emhp_merged", "adjusted_p_value"])`, r² = `r round(emhp_merged$r2, 4)`) (Figure 2). Cross-sectionally, hippocampal volume and memory were not significantly correlated at baseline (_p_ = `r format_p_value(model_baseline$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "model_baseline", "adjusted_p_value"])`) but were significantly correlated at follow-up (β = `r round(model_followup$coef, 4)`, _p_ = `r format_p_value(model_followup$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "model_followup", "adjusted_p_value"])`, r² = `r round(model_followup$r2, 4)`). These results are summarized in Table 4. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
# 1) Fit the lm model
lm_mem_hc_study <- lm(memory_slopes ~ hc_slopes + study + age_baseline + FU_time + sex + edu, data = clean_data)

# 2) Create a reference grid over a range of hc_slopes
rg_mem_hc <- ref_grid(
  lm_mem_hc_study,
  at = list(hc_slopes = seq(
    from = min(clean_data$hc_slopes),
    to = max(clean_data$hc_slopes),
    length.out = 50
  )),
)

# 3) Get the predicted means for that grid
em_means_mem_hc <- emmeans(rg_mem_hc, ~hc_slopes)
em_df_mem_hc <- as.data.frame(em_means_mem_hc)

# 4) Plot the average slope with confidence bands
mem_hc_plot <- ggplot(em_df_mem_hc, aes(x = hc_slopes, y = emmean)) +
  geom_point(
    data = clean_data,
    aes(x = hc_slopes, y = memory_slopes),
    alpha = 0.5
  ) +
  geom_ribbon(
    aes(ymin = lower.CL, ymax = upper.CL),
    fill = "#B4D4FF",
    alpha = 0.2
  ) +
  geom_line(
    size = 1.2,
    color = "#86B6F6"
  ) +
  labs(
    y = "Memory Annual Change",
    x = "Hippocampal Annual Change"
  ) +
  theme_minimal() +
  theme(
    legend.position = "right",
    plot.background = element_rect(fill = "white", color = "white"), # Set background to white
    plot.margin = margin(t = 0.35, r = 0.35, b = 0.35, l = 0.35, unit = "cm"),
    axis.text.x = element_text(margin = margin(t = -5)),
    plot.title = element_text(hjust = 0.5)
  )
```

__Figure 2:__ _Memory vs Hippocampal Annual Change_

```{r fig1, echo=FALSE, dpi = 400, fig.width=6, fig.height=4, message=FALSE, warning=FALSE}
mem_hc_plot
```

The association between memory and hippocampus annual change plotted using a linear model that controls for all covariates using estimated marginal means. The model produces a coefficient of `r format(round(emhp_merged_unscaled$coef, 4), scientific = FALSE)` (scaled: `r round(emhp_merged$coef, 4)`) with confidence intervals. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Create the table
table_data_emhc <- data.frame(
  " " = c(
    "", "**Memory Annual Change**"
  ),
  "**Hippocampus Annual Change**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(emhp_merged$coef, 4), " (", round(emhp_merged$ci[1], 4), " to ", round(emhp_merged$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "emhp_merged", "adjusted_p_value"])
  ),
  " " = c(
    "_Adjusted R²_", round(emhp_merged$adj_r2, 4)
  ),
  " " = c(
    "", "**Memory Baseline**"
  ),
  "**Hippocampus Baseline**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(model_baseline$coef, 4), " (", round(model_baseline$ci[1], 4), " to ", round(model_baseline$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "model_baseline", "adjusted_p_value"])
  ),
  " " = c(
    "_Adjusted R²_", round(model_baseline$adj_r2, 4)
  ),
  " " = c(
    "", "**Memory Follow-up**"
  ),
  "**Hippocampus Follow-up**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(model_followup$coef, 4), " (", round(model_followup$ci[1], 4), " to ", round(model_followup$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "model_followup", "adjusted_p_value"])
  ),
  " " = c(
    "_Adjusted R²_", round(model_followup$adj_r2, 4)
  )
)
```

__Table 4__: _Linear Regressions with Hippocampus Volume and Memory _

```{r tab4, echo=FALSE} 
# Format and print the table
kable(table_data_emhc, format = "markdown", col.names = c(" ", "**Hippocampus Annual Change**", " ", " ", " ", "**Hippocampus Baseline**", " ", " ", " ", "**Hippocampus Follow-up**", " ", " "))
```
Note. Abbreviations: CI Confidence Interval.

```{r echo=FALSE}
# Calculate the combined CR and BM value
clean_data <- clean_data %>%
  mutate(
    high_cr = ifelse(CR > 0.5, "High CR", "Low CR")
  )

# Calculate mean and SD for CR and BM variables
mean_bm <- mean(clean_data$BM, na.rm = TRUE)
min_bm <- min(clean_data$BM, na.rm = TRUE)
max_bm <- max(clean_data$BM, na.rm = TRUE)

mean_cr <- mean(clean_data$CR, na.rm = TRUE)
min_cr <- min(clean_data$CR, na.rm = TRUE)
max_cr <- max(clean_data$CR, na.rm = TRUE)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Run LME for sensitivity analysis 

# Pivot data from wide to long
long_clean_data <- clean_data %>%
  pivot_longer(
    cols = matches("(.+)\\.(\\d+)$"),             
    names_to = c("base_var", "timepoint"),
    names_pattern = "(.+)\\.(\\d+)$",
    values_to = "value"
  ) %>%
  mutate(timepoint = as.numeric(timepoint)) %>%   # Convert timepoint to numeric
  pivot_wider(
    names_from = "base_var",
    values_from = "value"
  )

# Add a function for getting the stats from LME models
get_mixed_effects_stats <- function(formula, data, vars_priority) {
  # Fit the mixed-effects model
  model <- lmer(formula, data = data, REML = FALSE) # uses Full Maximum Likelihood 
  summary_model <- summary(model)
  coefficients <- summary_model$coefficients
  anova_results <- anova(model) # Get ANOVA results for p-values

  # Initialize the variable index
  variable_index <- NULL

  # Check for each variable in priority order and break on the first match
  for (var in vars_priority) {
    if (var %in% rownames(coefficients)) {
      variable_index <- which(rownames(coefficients) == var)
      break
    }
  }

  # Check if a relevant variable was found
  if (is.null(variable_index)) {
    stop("None of the priority variables are in the model.")
  }

  coef <- coefficients[variable_index, 1]
  ci <- confint(model, parm = rownames(coefficients)[variable_index])

  # Extract p-value from ANOVA table
  p_value <- anova_results[which(rownames(anova_results) == rownames(coefficients)[variable_index]), "Pr(>F)"]
  r2 <- as.data.frame(VarCorr(model))$vcov[1]

  # Compute marginal and conditional R² 
  r2_vals <- r.squaredGLMM(model)
  marginal_r2 <- r2_vals[1]       # Variance explained by fixed effects only
  conditional_r2 <- r2_vals[2]    # Variance explained by fixed + random effects
  
  return(list(
    coef = coef,
    ci = ci,
    p_value = p_value,
    marginal_r2 = marginal_r2,
    conditional_r2 = conditional_r2
  ))
}

# Run Spearman's correlation for CRBM 
crbm_corr <- cor.test(clean_data$CR, clean_data$BM, method = "spearman")
crbm_corr$p_value <- crbm_corr$p.value
```

__**Analyzing Cognitive Reserve and Brain Maintenance Variables:**__

We established CR and BM variables as illustrated in Figure 3, which plots each participant’s hippocampal and memory annual change - both adjusted for age - together with their corresponding CR or BM value. Because these trajectories are adjusted for age, individuals whose hippocampus atrophies and memory declines less than expected for their age appear to show increases in both hippocampal volume and memory. This adjustment for age allows the CR and BM variables to consider the faster rate of decline in hippocampal volume and memory observed among our older participants (Supplementary Figure 1).

```{r echo=FALSE, warning=FALSE}
custom_colors_bm <- c("#E5E483", "#8ac926", "#4CCD99", "#5BBCFF", "#024CAA", "#00224D")
custom_colors_cr <- c("#f5d7b0", "#f5be27", "#fa8d00", "#bc507f", "#7E1891", "#1c0159")

cr_plot <- ggplot(clean_data, aes(y = res_mem_slopes_age, x = res_hc_slopes_age, color = (CR))) +
  geom_point(size = 2, alpha = 0.7) +
  scale_color_gradientn(
    colors = custom_colors_cr,
    limits = c(0, 1),
    values = c(0, 1 / 3, 2 / 3, 1) # Values mapped to scale so both figures have the same scale
  ) +
  geom_abline(slope = 0, intercept = 0, color = "grey", size = 1, alpha = 0.7) +
  geom_abline(slope = 10000, intercept = 0, color = "grey", size = 1, alpha = 0.7) +
  labs(
    y = "Memory Change (age-adjusted)",
    x = "Hippocampus Change (age-adjusted)",
    color = "Cognitive Reserve",
    tag = "B)"
  ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        plot.tag = element_text(size = 12), 
        plot.tag.position = c(0.08, 0.08)) +
  ylim(-0.6, 0.9) +
  xlim(-400, 250)

bm_plot <- ggplot(clean_data, aes(y = res_mem_slopes_age, x = res_hc_slopes_age, color = BM)) +
  geom_point(size = 2, alpha = 0.7) +
  scale_color_gradientn(
    colors = custom_colors_bm,
    limits = c(0, 1),
    values = c(0, 1 / 3, 2 / 3, 1) # Values mapped to scale so both figures have the same scale
  ) +
  geom_abline(slope = 0, intercept = 0, color = "grey", size = 1, alpha = 0.7) +
  geom_abline(slope = 10000, intercept = 0, color = "grey", size = 1, alpha = 0.7) +
  labs(
    y = "Memory Change (age-adjusted)",
    x = "Hippocampus Change (age-adjusted)",
    color = "Brain Maintenance",
    tag = "A)"
  ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        plot.tag = element_text(size = 12), 
        plot.tag.position = c(0.08, 0.08)) +
  ylim(-0.6, 0.9) +
  xlim(-400, 250)

# Combine the plots into one figure
combined_crbm <- bm_plot + cr_plot + plot_layout(ncol = 2)
```

__Figure 3:__ _Brain Maintenance and Cognitive Reserve Plots_

```{r fig2, echo=FALSE, dpi = 400, fig.width=7, fig.height=4, message=FALSE, warning=FALSE}
combined_crbm
```
<small>A) The association between hippocampus and memory annual change, adjusted for age. The color code represents the BM value for a given participant. B) The same association between hippocampus and memory annual change, adjusted for age. The color code represents the CR value for a given participant. Abbreviations: CR Cognitive Reserve, BM Brain Maintenance.</small>

The mean BM value was  `r round(mean_bm, 4)` (range = `r round(min_bm, 4)` - `r round(max_bm, 4)`), and the mean CR value was `r round(mean_cr, 4)` (range = `r round(min_cr, 4)` - `r round(max_cr, 4)`). To visualize how CR moderates the hippocampal-memory relationship, we examined the association between hippocampal and memory change across increasing levels of CR (Figure 4). We divided participants into CR bins and estimating local linear fits within each bin to visualize a continuous moderation effect. The slope of the hippocampal-memory relationship decreased as CR increased, indicating that individuals with higher CR showed weaker coupling between hippocampal decline and memory decline. We found a small positive association between our CR and BM variables (Spearman’s ρ = `r round(crbm_corr$estimate, 4)`, _p_ = `r signif(crbm_corr$p.value, 4)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "crbm_corr", "adjusted_p_value"])`). Sensitivity analyses showed that alternative formulations of the CR and BM variables were highly correlated with the main definitions (Spearman’s ρ = 0.74–1.00, _ps_ = < 0.0001; Supplementary Table 2, Figure 2). 

 
__Figure 4:__ _Change in the Memory–Hippocampal Relationship Across Levels of Cognitive Reserve_


The relationship between hippocampal and memory change, both adjusted for age, shown as a function of CR. Data are divided into quantile bins of CR, and a linear fit between hippocampal and memory change for each bin is displayed as a line segment. Line color indicates the mean CR value for each bin. The figure illustrates that the slope of the memory–hippocampal association varies across levels of CR, suggesting that individuals with higher CR show less coupling between hippocampal and memory change.

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Education and CR
vars <- c(
  "scale(edu):scale(adj_HIP_total)",
  "scale(edu)"
)

int_vars <- c(
  "scale(CR)"
)

# Check if edu is a CR factor
edu_cr_int <- get_mixed_effects_stats(scale(memory) ~ scale(edu) * scale(adj_HIP_total) + (1|id) + (1|study) + age + FU_time + sex, data = long_clean_data, vars)

# Then edu and memory
edu_mem <- get_mixed_effects_stats(scale(memory) ~ scale(edu) + (1|id) + (1|study) + age + FU_time + sex, data = long_clean_data, vars)

# Then is edu related to our CR variable
cr_pathway <- clean_data %>%
  filter(high_cr == "High CR") 
  
edu_cr <- get_regression_stats(scale(edu) ~ scale(CR) + study + age_baseline + sex, data = cr_pathway, int_vars)

# Check then if explains mem above and beyond HC
edu_cr_hc <- get_mixed_effects_stats(scale(memory) ~ scale(edu) + scale(adj_HIP_total) + (1|id) + (1|study) + age + FU_time + sex, data = long_clean_data, vars)
```

__**Cognitive Reserve, Education, and Functional Connectivity:**__

Longitudinal mixed-effects analyses provided no evidence that education moderated the relationship between hippocampal volume and memory (_p_ = `r format_p_value(edu_cr_int$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "edu_cr_int", "adjusted_p_value"])`), nor was there a significant relationship between our CR variable and education (_p_ = `r format_p_value(edu_cr$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "edu_cr", "adjusted_p_value"])`) in high CR participants. However, education was significantly correlated with memory (β = `r round(edu_mem$coef, 4)`, _p_ = `r format(format_p_value(edu_mem$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "edu_mem", "adjusted_p_value"])`, marginal r² = `r round(edu_mem$marginal_r2, 4)`, conditional r² = `r round(edu_mem$conditional_r2, 4)`), and the correlation remained significant when hippocampal change was included in the model (β = `r round(edu_cr_hc$coef, 4)`, _p_ = `r format(format_p_value(edu_cr_hc$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "edu_cr_hc", "adjusted_p_value"])`, marginal r² = `r round(edu_cr_hc$marginal_r2, 4)`, conditional r² = `r round(edu_cr_hc$conditional_r2, 4)`; Table 5). 

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Set the priority variables
priority_var_cr <- c(
  "scale(dmn):scale(adj_HIP_total)",
  "scale(ec):scale(adj_HIP_total)",  
  "scale(sn):scale(adj_HIP_total)",
  "scale(dmn_baseline):scale(adj_HIP_total_baseline)",
  "scale(dmn_followup):scale(adj_HIP_total_followup)",
  "scale(ec_baseline):scale(adj_HIP_total_baseline)",
  "scale(ec_followup):scale(adj_HIP_total_followup)",
  "scale(sn_baseline):scale(adj_HIP_total_baseline)",
  "scale(sn_followup):scale(adj_HIP_total_followup)"
)

cr_dmn_stats <- get_mixed_effects_stats(scale(memory) ~ scale(dmn) * scale(adj_HIP_total) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_cr)
cr_ecn_stats <- get_mixed_effects_stats(scale(memory) ~ scale(ec) * scale(adj_HIP_total) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_cr)
cr_sn_stats <- get_mixed_effects_stats(scale(memory) ~ scale(sn) * scale(adj_HIP_total) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_cr)
```


```{r echo=FALSE, message=FALSE, warning=FALSE}
# Run the interaction model separately 
priority_var_int_cr <- c(
  "scale(CR)",
  "scale(ec)",
  "scale(dmn)",
  "scale(sn)",
  "CR"
)

# Filter to those aging CR-like
long_cr_pathway_fc <- long_clean_data %>%
  filter(high_cr == "High CR") 

cr_dmn <- get_mixed_effects_stats(scale(dmn) ~ scale(CR) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_cr_pathway_fc, priority_var_int_cr)
cr_ec <- get_mixed_effects_stats(scale(ec) ~ scale(CR) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_cr_pathway_fc, priority_var_int_cr)
cr_sn <- get_mixed_effects_stats(scale(sn) ~ scale(CR) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_cr_pathway_fc, priority_var_int_cr)

cr_dmn_mem <- get_mixed_effects_stats(scale(memory) ~ scale(dmn) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_int_cr)
cr_ec_mem <- get_mixed_effects_stats(scale(memory) ~ scale(ec) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_int_cr)
cr_sn_mem <- get_mixed_effects_stats(scale(memory) ~ scale(sn) + (1|id) + (1|study) + age + FU_time + sex + edu, data = long_clean_data, priority_var_int_cr)
```


With functional connectivity, the longitudinal mixed-effects analyses revealed that anterior SN connectivity moderated the relationship between hippocampal volume and memory (β = `r round(cr_sn_stats$coef, 4)`, _p_ = `r format(format_p_value(cr_sn_stats$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_sn_stats", "adjusted_p_value"])`, marginal r² = `r round(cr_sn_stats$marginal_r2, 4)`, conditional r² = `r round(cr_sn_stats$conditional_r2, 4)`), such that participants with higher anterior SN connectivity did not have a significant relationship between memory and hippocampal volume. Anterior SN connectivity was also positively correlated with memory (β = `r round(cr_sn_mem$coef, 4)`, _p_ = `r format(format_p_value(cr_sn_mem$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_sn_mem", "adjusted_p_value"])`, marginal r² = `r round(cr_sn_mem$marginal_r2, 4)`, conditional r² = `r round(cr_sn_mem$conditional_r2, 4)`), and there was a positive relationship between our CR variable and anterior SN connectivity in high CR participants (β = `r round(cr_sn$coef, 4)`, _p_ = `r format(format_p_value(cr_sn$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_sn", "adjusted_p_value"])`, marginal r² = `r round(cr_sn$marginal_r2, 4)`, conditional r² = `r round(cr_sn$conditional_r2, 4)`). There was also evidence that DMN connectivity moderated the hippocampal volume-memory relationship (β = `r round(cr_dmn_stats$coef, 4)`, _p_ = `r format_p_value(cr_dmn_stats$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_dmn_stats", "adjusted_p_value"])`, marginal r² = `r round(cr_dmn_stats$marginal_r2, 4)`, conditional r² = `r round(cr_dmn_stats$conditional_r2, 4)`); however, unlike with anterior SN connectivity, when DMN connectivity is high, there is a significant relationship between hippocampal and memory change, such that decreases in hippocampal volume are associated with increases in memory. The moderation analyses are summarized in Figure 5 which depicts how the hippocampal volume-memory relationship increases as anterior SN connectivity decreases, demonstrating a significant positive relationship between memory and hippocampal volume only for participants with anterior SN connectivity less than -0.02SD above the mean. Figure 5 also depicts how the hippocampal volume-memory relationship increases as DMN connectivity decreases; however, in this case, it demonstrates a significant positive relationship between memory and hippocampal volume for participants with DMN connectivity less than 0.08SD above the mean and for participants with DMN connectivity greater than 3.7SD above the mean. This indicates that, among participants with high DMN connectivity, there is a negative hippocampal-memory relationship. Additionally with DMN connectivity, there was no evidence of a relationship between DMN connectivity and memory (_p_ = `r format_p_value(cr_dmn_mem$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_dmn_mem", "adjusted_p_value"])`), or of a relationship between our CR variable and DMN connectivity in high CR participants (_p_ = `r format_p_value(cr_dmn$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_dmn", "adjusted_p_value"])`).

With ECN connectivity, there was no evidence that it moderated the relationship between hippocampal volume and memory (_p_ = `r format_p_value(cr_ecn_stats$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_ecn_stats", "adjusted_p_value"])`). There was also no evidence of a relationship between ECN connectivity and memory (_p_ = `r format_p_value(cr_ec_mem$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_ec_mem", "adjusted_p_value"])`), or of a relationship between our CR variable and ECN connectivity in high CR participants (_p_ = `r format_p_value(cr_ec$p_value)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "cr_ec", "adjusted_p_value"])`). All functional connectivity results are summarized in Table 5. Figure 6 shows the difference in anterior SN connectivity when comparing low versus high CR groups. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Fit the model 
model <- lmer(scale(memory) ~ scale(sn) * scale(adj_HIP_total) + (1|id) + 
              study + age + FU_time + sex + edu, 
              data = long_clean_data)

# Johnson-Neyman plot
jn_result <- johnson_neyman(model, 
                           pred = "scale(adj_HIP_total)", 
                           modx = "scale(sn)",
                           plot = TRUE)
jn_plot <- jn_result$plot  
jn_plot <- jn_plot + 
  labs(title = NULL, tag = "A)") +
  xlab("Anterior Salience Network Connectivity") +
  ylab("Effect of Hippocampal Volume on Memory") +
  labs(title = "Johnson-Neyman Plot") +
  theme(plot.caption = element_text(size = 10, hjust = -0.2),
        legend.position = "bottom",
        plot.tag.position = c(1, 0.01),  # bottom-left
        title = NULL,
        axis.title.x = element_text(size = 11),
        axis.title.y = element_text(size = 11),
        plot.title = element_text(face ="plain", size = 11)
        )

cr_fc <- jn_plot + 
  xlab("Anterior Salience Network Connectivity") +
  ylab("Effect of Hippocampal Volume on Memory") 

# Fit 2nd model 
model_dmn <- lmer(scale(memory) ~ scale(dmn) * scale(adj_HIP_total) + (1|id) + 
              study + age + FU_time + sex + edu, 
              data = long_clean_data)

# Johnson-Neyman plot
jn_plot_dmn <- jn_result_dmn$plot +
  labs(title = NULL, tag = "B)") +
  xlab("Default Mode Network Connectivity") +
  ylab("Effect of Hippocampal Volume on Memory") +
  theme(
    legend.position = "bottom",
    plot.tag.position = c(0, 0.01),  # bottom-left
    plot.tag = element_text(size = 12, face = "bold"),
    plot.caption = element_text(size = 10, hjust = -0.2),
    axis.title.x = element_text(size = 11),
    axis.title.y = element_text(size = 11),
    plot.title = element_text(face = "plain", size = 11)
  )

# Remove the significance shading bar manually
jn_plot_dmn$layers <- jn_plot_dmn$layers[-1]

cr_fc_dmn <- jn_plot_dmn + 
  xlab("Default Mode Network Connectivity") +
  ylab("Effect of Hippocampal Volume on Memory") 

# Combine the plots and collect legend
cr_fc <- cr_fc + cr_fc_dmn + plot_layout(ncol = 2, guides = "collect") &
  theme(legend.position = "bottom")
```

__Figure 5:__ _Anterior Salience Network Functional Connectivity and the Hippocampal-Memory Relationship_

```{r fig5, echo=FALSE, dpi = 400, fig.width=7, fig.height=5, message=FALSE, warning=FALSE}
cr_fc
```
A) Johnson-Neyman plot of how the hippocampal volume-memory relationship changes as anterior Salience Network connectivity changes. The plot shows that as anterior Salience Network connectivity decreases, the hippocampal volume-memory relationship increases, only becoming significant when Salience Network connectivity is less than -0.02SD above the mean. Above that threshold, the hippocampal volume–memory relationship is not significant. B) The hippocampal volume-memory this time depicted with Default Mode Network connectivity changes. The plot shows that as Default Mode Network connectivity decreases, the hippocampal volume-memory relationship increases. However, this hippocampal-memory relationship is significant for both high and low Default Mode Network connectivity values, only becoming non-significant when Default Mode Network connectivity is between 0.08-3.70SD above the mean. Within that range, the hippocampal volume–memory relationship is not significant. Abbreviations: SN Salience Network, DMN Default Mode Network, n.s. not significant. 

__Figure 6:__ _Cognitive Reserve and Anterior Salience Network Connectivity_

Connectivity differences (high CR vs low CR) within the anterior SN among participants classified as aging CR-like. For each region, connectivity was averaged across sessions, scaled, adjusted for scanner/cohort effects, and expressed as the difference between the high CR and low CR groups; all values are positive, indicating stronger connectivity in the high CR group in all regions. Abbreviations: CR, cognitive reserve; SN, Salience Network.    

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Create the table
long_fc <- data.frame(
  " " = c(
    "", "**Memory**",
    "", "", "**Memory**", 
    "", "", "**Cognitive Reserve**"
  ),
  "**Education * Hippocampal Volume**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(edu_cr_int$coef, 4), " (", round(edu_cr_int$ci[1], 4), " to ", round(edu_cr_int$ci[2], 4), ")"),
    "**Education**", "_Coefficient (95% CI)_",
    paste0(round(edu_mem$coef, 4), " (", round(edu_mem$ci[1], 4), " to ", round(edu_mem$ci[2], 4), ")"),
    "**Education**", "_Coefficient (95% CI)_",
    paste0(round(edu_cr$coef, 4), " (", round(edu_cr$ci[1], 4), " to ", round(edu_cr$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "edu_cr_int", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "edu_mem", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "edu_cr", "adjusted_p_value"])
    ),
  " " = c(
    "_Marginal R²_", round(edu_cr_int$marginal_r2, 4),
    "", "_Marginal R²_", round(edu_mem$marginal_r2, 4),
    "", "_Adjusted R²_", round(edu_cr$adj_r2, 4)
  ),
  "**DMN Connectivity * Hippocampal Volume**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(cr_dmn_stats$coef, 4), " (", round(cr_dmn_stats$ci[1], 4), " to ", round(cr_dmn_stats$ci[2], 4), ")"),
    "**DMN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_dmn_mem$coef, 4), " (", round(cr_dmn_mem$ci[1], 4), " to ", round(cr_dmn_mem$ci[2], 4), ")"),
    "**DMN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_dmn$coef, 4), " (", round(cr_dmn$ci[1], 4), " to ", round(cr_dmn$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_dmn_stats", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_dmn_mem", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_dmn", "adjusted_p_value"])
  ),
  " " = c(
    "_Marginal R²_", round(cr_dmn_stats$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_dmn_mem$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_dmn$marginal_r2, 4)
  ),
  "**ECN Connectivity * Hippocampal Volume**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(cr_ecn_stats$coef, 4), " (", round(cr_ecn_stats$ci[1], 4), " to ", round(cr_ecn_stats$ci[2], 4), ")"),
    "**ECN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_ec_mem$coef, 4), " (", round(cr_ec_mem$ci[1], 4), " to ", round(cr_ec_mem$ci[2], 4), ")"),
    "**ECN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_ec$coef, 4), " (", round(cr_ec$ci[1], 4), " to ", round(cr_ec$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_ecn_stats", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_ec_mem", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_ec", "adjusted_p_value"])
  ),
  " " = c(
    "_Marginal R²_", round(cr_ecn_stats$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_ec_mem$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_ec$marginal_r2, 4)
  ),
  "**Anterior SN Connectivity * Hippocampal Volume**" = c(
    "_Coefficient (95% CI)_",
    paste0(round(cr_sn_stats$coef, 4), " (", round(cr_sn_stats$ci[1], 4), " to ", round(cr_sn_stats$ci[2], 4), ")"),
    "**Anterior SN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_sn_mem$coef, 4), " (", round(cr_sn_mem$ci[1], 4), " to ", round(cr_sn_mem$ci[2], 4), ")"),
    "**Anterior SN Connectivity**", "_Coefficient (95% CI)_",
    paste0(round(cr_sn$coef, 4), " (", round(cr_sn$ci[1], 4), " to ", round(cr_sn$ci[2], 4), ")")
  ),
  " " = c(
    "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_sn_stats", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_sn_mem", "adjusted_p_value"]),
    "", "_Adjusted p-value_", format_p_value(p_value_df[p_value_df$model_name == "cr_sn", "adjusted_p_value"])
  ),
  " " = c(
    "_Marginal R²_", round(cr_sn_stats$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_sn_mem$marginal_r2, 4),
    "", "_Marginal R²_", round(cr_sn$marginal_r2, 4)
  )
)
```

__Table 5__: _Models Examining Education, Functional Connectivity and Cognitive Reserve_
```{r tab5, echo=FALSE} 
# Format and print the table
kable(long_fc, format = "markdown", col.names = c(" ", "**Education * Hippocampal Volume**", " ", " ", "**DMN Connectivity * Hippocampal Volume**", " ", " ", "**ECN Connectivity * Hippocampal Volume**", " ", " ", "**Anterior SN Connectivity * Hippocampal Volume**", " ", " "))
```

Note. Abbreviations: DMN Default Mode Network, ECN Executive Control Network, SN Salience Network, CI Confidence Interval.

**Supplementary Materials**

Supplementary Figure 1 depicts distributions of normalized memory composite scores by cohort across timepoints. The box-and-jitter plots show the distribution of harmonized memory scores across cohorts following the normalization process. The resulting distributions overlap across cohorts, with means centered near zero and standard deviations close to one, indicating that the harmonization procedure places all cohorts on a common scale. 

**_Cohort Distributions of Episodic Memory Scores_**
```{r echo=FALSE, message=FALSE, warning=FALSE}
# Function to plot the memory scores by cohort
plot_memory_all_tps <- function(df, cohort, tps) {
  cohort <- enquo(cohort)
  tp_names <- tidyselect::eval_select(enquo(tps), df) |> names()

  long_tp <- df %>%
    select(!!cohort, all_of(tp_names)) %>%
    pivot_longer(all_of(tp_names), names_to = "timepoint", values_to = "score") %>%
    filter(!is.na(score)) %>%
    mutate(timepoint = dplyr::recode(timepoint,
                                     "memory.1" = "Timepoint 1",
                                     "memory.2" = "Timepoint 2",
                                     "memory.3" = "Timepoint 3",
                                     "memory.4" = "Timepoint 4",
                                     "memory.5" = "Timepoint 5",
                                     "memory.6" = "Timepoint 6"))
  long_tp <- long_tp %>%
    mutate(!!quo_name(cohort) := forcats::fct_recode(
      !!cohort,
      "Barcelona WAHA" = "WAHA",
      "Oslo LCBC"      = "Oslo",
      "Umea Betula"    = "Betula",
      "Umea COBRA"     = "Cobra"
    ))

  
  # Compute mean and SD for annotation
  mean_labels <- long_tp %>%
    group_by(!!cohort, timepoint) %>%
    summarise(
      mean = mean(score, na.rm = TRUE),
      sd = sd(score, na.rm = TRUE),
      n = n(),
      .groups = "drop"
    ) %>%
    mutate(
      label = sprintf("M=%.2f\nSD=%.2f\nn=%d", mean, sd, n)
    )
  
  p <- ggplot(long_tp, aes(x = !!cohort, y = score, fill = !!cohort)) +
    geom_violin(trim = FALSE, width = 0.9, alpha = 0.35, color = NA) +     
    geom_boxplot(width = 0.45, outlier.shape = NA, alpha = 0.9,           
                 color = "black") +
    geom_jitter(width = 0.12, alpha = 0.45, size = 1.2) +                  
    facet_wrap(~ timepoint, ncol = 3, scales = "fixed") +                
    geom_hline(yintercept = 0, linetype = "dashed", linewidth = 0.3, alpha = 0.8) +
    geom_text(
      data = mean_labels,
      aes(x = !!cohort, y = 2.5, label = label),
      size = 3, color = "black"
    ) +
    labs(x = NULL, y = "Memory Composite", fill = "Cohort") +
    theme_minimal(base_size = 12) +
    theme(
      legend.position = "top",
      axis.text.x = element_text(angle = 20, hjust = 1),
      panel.grid.major.x = element_blank()
    ) +
    coord_cartesian(ylim = c(-3, 4)) +
    scale_y_continuous(expand = expansion(mult = c(0.05, 0.1))) +          
    scale_fill_manual(values = c(
      "Barcelona WAHA" = "indianred2",
      "Oslo LCBC"      = "steelblue2",
      "Umea Betula"    = "#86db49",
      "Umea COBRA"     = "#F7CA18"
    )) 

  return(p)
}

plot_memory_all_tps(
  clean_data,
  cohort = cohort,
  tps    = c(memory.1, memory.2, memory.3, memory.4, memory.5, memory.6)
)
```

Note. Abbreviations: M mean, SD standard deviation, n number.

**_Linear Mixed Effects and Generalized Additive Mixed Models for Memory, Hippocampus Volume, and Age_**
```{r echo=FALSE, message=FALSE, warning=FALSE}
# Set a list of priority variables
vars <- c(
  "scale(age):scale(CR)",
  "scale(age):scale(BM)"
)

# Linear model for comparison with the gam model
mod_lme_mem <- lmer(memory ~ age + (1|id), data = long_clean_data)
aic_lm_mem <- AIC(mod_lme_mem)
bic_lm_mem <- BIC(mod_lme_mem)

mod_lme_hc <- lmer(adj_HIP_total ~ age + (1|id), data = long_clean_data)
aic_lm_hc <- AIC(mod_lme_hc)
bic_lm_hc <- BIC(mod_lme_hc)

# Get the regression stats
age <- c("scale(age)")

suppressMessages({
  mem_age_lme <- get_mixed_effects_stats(scale(memory) ~ scale(age) + (1|id), data = long_clean_data, age)
  hc_age_lme <- get_mixed_effects_stats(scale(adj_HIP_total) ~ scale(age) + (1|id), data = long_clean_data, age)
})
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Uncomment this code to evaluate and check GAMM assumptions

# # Loop through k values and print results
# for (k in 1:27) {
#   model_mem <- gamm4(memory ~ s(age, k = k+3), data = long_clean_data, random = ~(1|id))
# 
#   cat("Model with k =", k + 3, "\n")
#   cat("AIC:", AIC(model_mem$mer), "\n")
#   cat("BIC:", BIC(model_mem$mer), "\n")
#   cat("LogLik:", logLik(model_mem$mer), "\n")
#   cat("\n")
# }
# 
# # Loop through k values and print results
# for (k in 1:27) {
#   model_hc <- gamm4(adj_HIP_total ~ s(age, k = k+3), data = long_clean_data, random = ~(1|id))
# 
#   cat("Model with k =", k + 3, "\n")
#   cat("AIC:", AIC(model_hc$mer), "\n")
#   cat("BIC:", BIC(model_hc$mer), "\n")
#   cat("LogLik:", logLik(model_hc$mer), "\n")
#   cat("\n")
# }
# 
# gamm_memory <- gamm4(memory ~ s(age, k = 4), data = long_clean_data, random = ~(1|id))
# gam.check(gamm_memory$gam)
# 
# gamm_hc <- gamm4(adj_HIP_total ~ s(age, k = 5), data = long_clean_data, random = ~(1|id))
# gam.check(gamm_hc$gam)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Function to calculate and store gamm model metrics
get_gamm_metrics <- function(k, df, response_var, lme_mod) {
  max_k <- min(k, length(unique(df$age)) - 1)

  if (max_k > 1) {
    # Dynamically use the response variable and correct random effect
    mod_em <- gamm4(as.formula(paste(response_var, "~ 1 + s(age, k =", max_k, ")")),
      data = df, random = ~ (1 | id), REML   = FALSE
    )

    # Calculate required metrics
    aic <- AIC(mod_em$mer)
    bic <- BIC(mod_em$mer)
    loglik <- logLik(mod_em$mer)

    # Extract EDF
    edf <- sum(summary(mod_em$gam)$edf)

    # Calculate p-value for the GAMM terms
    gamm_summary <- summary(mod_em$gam)
    p_value_gamm <- gamm_summary$s.pv

    # Likelihood ratio test
    lrt_stat <- -2 * (logLik(lme_mod) - loglik) # Comparing the lme to the gamm model
    lrt_df <- df.residual(lme_mod) - df.residual(mod_em$mer)
    p_value_lrt <- 1 - pchisq(lrt_stat, lrt_df)

    # Return a named list of metrics
    return(data.frame(
      k = max_k, AIC = aic, BIC = bic, LogLik = loglik,
      EDF = edf, LRT_df = lrt_df, "LRT test stat" = lrt_stat,
      "p_value_gamm" = as.numeric(p_value_gamm),
      "p_value_lrt" = as.numeric(p_value_lrt)
    ))
  } else {
    return(data.frame(
      k = max_k, AIC = NA, BIC = NA,
      LogLik = NA, EDF = NA, "LRT test stat" = NA,
      "p_value_gamm" = NA, "p_value_lrt" = NA
    ))
  }
}

# Function to fit and compare models over k range for different response variables
gamm_compare <- function(response_var, mod_lme) {
  compare_gamm <- do.call(rbind, lapply(k_range, function(k) {
    get_gamm_metrics(k, long_clean_data, response_var, mod_lme)
  }))
  return(compare_gamm)
}

# Run the comparison for memory and adjusted HIP
k_range <- 4 # Define k range for memory
gamm_compare_em_memory <- gamm_compare("memory", mod_lme_mem)

k_range <- 5 # Define k range for HC
gamm_compare_em_adj_hc <- gamm_compare("adj_HIP_total", mod_lme_hc)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
gamm_table <- data.frame(
  " " = c(
    "",
    "**GAMM**",
    "**LME Model**"
  ),
  "**Memory**" = c(
    "_AIC_",
    paste0(round(gamm_compare_em_memory$AIC, 4)),
    paste0(round(aic_lm_mem, 4))
  ),
  " " = c(
    "_BIC_",
    paste0(round(gamm_compare_em_memory$BIC, 4)),
    paste0(round(bic_lm_mem, 4))
  ),
  " " = c(
    "_Adjusted LRT p-value_",
    format_p_value(p_value_df[p_value_df$model_name == "gamm_memory_lft", "adjusted_p_value"]),
    " "
  ),
  "**Hippocampus Volume**" = c(
    "_AIC_",
    paste0(round(gamm_compare_em_adj_hc$AIC, 4)),
    paste0(round(aic_lm_hc, 4))
  ),
  " " = c(
    "_BIC_",
    paste0(round(gamm_compare_em_adj_hc$BIC, 4)),
    paste0(round(bic_lm_hc, 4))
  ),
  " " = c(
    "_Adjusted LRT p-value_",
    format_p_value(p_value_df[p_value_df$model_name == "gamm_hc_lft", "adjusted_p_value"]),
    " "
  )
)
```

We initially fit the GAMMs using a range of basis functions and compared their fit to the LME model. Despite variations in the GAMM fits, the conclusions regarding how the GAMM and the LME models compared did not change. In Supplementary Table 1, we fit the GAMMs with four basis functions for the memory model and five basis functions for the hippocampus model as increasing the number beyond this point did not substantially change the AIC or BIC values. The smooth age-term was significant in the memory GAMM (effective degrees of freedom (EDF) = `r round(gamm_compare_em_memory$EDF, 4)`, _p_-value = `r format_p_value(gamm_compare_em_memory$p_value_gamm)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "gamm_memory_gamm", "adjusted_p_value"])`) and the hippocampus GAMM (EDF = `r round(gamm_compare_em_adj_hc$EDF, 4)`, _p_-value = `r format_p_value(gamm_compare_em_adj_hc$p_value_gamm)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "gamm_hc_gamm", "adjusted_p_value"])`). For the LME models, both models had significant age coefficients (memory: β = `r round(mem_age_lme$coef, 4)`, _p_ = `r format(format_p_value(mem_age_lme$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "mem_age_lme", "adjusted_p_value"])`; hippocampal volume β = `r round(hc_age_lme$coef, 4)`, _p_ = `r format(format_p_value(hc_age_lme$p_value), scientific = FALSE)`, _p_<sub>FDR</sub> = `r format_p_value(p_value_df[p_value_df$model_name == "hc_age_lme", "adjusted_p_value"])`). For memory, the LME and GAMM fits were similar; however, the GAMM fit was better according to AIC, BIC, and LRT. For hippocampal volume, the GAMM fit was better according to AIC, BIC, and LRT.

__Supplementary Table 1__: _Linear Mixed Effects and Generalized Additive Mixed Models_
```{r suptab1, echo=FALSE} 
# Format and print the table
kable(gamm_table, format = "markdown", col.names = c(" ", "**Memory**", " ", " ", "**Hippocampus**", " ", " "))
```
Note. Abbreviations: GAMM Generalized Additive Mixed Model, LME Linear Mixed Effects, AIC Akaike Information Criterion, BIC Bayesian Information Criterion, LRT Likelihood Ratio Test.

__Supplementary Figure 1__: _Memory and Hippocampus Volume Change with Age_
```{r echo=FALSE, message=FALSE, warning=FALSE}
# Define colors
color_palette <- c("indianred2", "steelblue2", "#86db49", "#F7CA18", "blue", "black")

# Create the plot age v memory
mem_age <- ggplot(long_clean_data, aes(x = age, y = memory, color = Study, group = id)) +
  geom_line(size = 0.4) +
  geom_point(size = 0.6) +
  scale_color_manual(values = color_palette) +
  geom_smooth(method = "gam", formula = y ~ s(x, k = 24, bs = "cs"), aes(group = 1), color = "grey", se = TRUE) +
  labs(x = "Age", y = "Episodic Memory", color = "Cohort", tag = "A)") +
  theme_minimal() +
  theme(legend.position = "bottom",
        plot.tag = element_text(size = 11), 
        plot.tag.position = c(0, 1))

# Create the plot age v HC
hc_age <- ggplot(long_clean_data, aes(x = age, y = adj_HIP_total, color = Study, group = id)) +
  geom_line(size = 0.4) +
  geom_point(size = 0.6) +
  scale_color_manual(values = color_palette) +
  geom_smooth(method = "gam", formula = y ~ s(x, k = 24, bs = "cs"), aes(group = 1), color = "grey", se = TRUE) +
  labs(x = "Age", y = "Hippocampal Volume", color = "Cohort", tag = "B)") +
  theme_minimal() +
  theme(legend.position = "none",
        plot.tag = element_text(size = 11), 
        plot.tag.position = c(0, 1)) # Remove legend for the second plot

# Combine the plots and collect legend
hc_mem_age <- mem_age + hc_age + plot_layout(ncol = 2, guides = "collect") &
  theme(legend.position = "bottom")
```

```{r supfig1, echo=FALSE, dpi = 400, fig.width=7, fig.height=4, message=FALSE, warning=FALSE}
hc_mem_age
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
vars_unscale <- c(
  "age"
)

mem_age_unscaled <- get_mixed_effects_stats((memory) ~ age + (1 | id), data = long_clean_data, vars_unscale)
```
<small>A) The association between memory and age with individual trajectories and a generalized additive mixed model (EDF = `r round(gamm_compare_em_memory$EDF, 4)`) and confidence intervals. B) The association between hippocampal volume and age with individual trajectories with a generalized additive mixed model (EDF = `r round(gamm_compare_em_adj_hc$EDF, 4)`) and confidence intervals. Abbreviations: EDF effective degrees of freedom.</small>

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Correcting for p-values

# Cohort comparisons 
cohort_comp <- list(
  age_model, age_model_fu, chisq_test, edu_model, 
  mmse_model, fu_time_model, fu_obs_model
)

# Memory, Hippocampus and Analyses
memory_hippocampus_edu_models <- list(emhp_merged, model_baseline, model_followup)

# Education and CR/BM Analyses
education_cr_models <- list(edu_cr, edu_cr_int, edu_mem, edu_cr_hc)

# CR and BM Sensitivity Analyses
cr_sensitivity_models <- list(crbm_corr)

# ECN and CR Analyses Longitudinal
fc_long_cr_ecn <- list(cr_ecn_stats, cr_ec, cr_ec_mem)

# DMN and CR Analyses Longitudinal
fc_long_cr_dmn <- list(cr_dmn_stats, cr_dmn, cr_dmn_mem)

# SN and CR Analyses Longitudinal
fc_long_cr_sn <- list(cr_sn_stats, cr_sn, cr_sn_mem)

# Duplicate the GAMMs to get separate p-values
gamm_memory_lft <- gamm_compare_em_memory %>% select(p_value_lrt)

gamm_memory_gamm <- gamm_compare_em_memory %>% select(p_value_gamm)

gamm_hc_lft <- gamm_compare_em_adj_hc %>% select(p_value_lrt)

gamm_hc_gamm <- gamm_compare_em_adj_hc %>% select(p_value_gamm)

# Model Comparisons and General Analyses
model_comp_general_models <- list(
  gamm_memory_lft, gamm_memory_gamm,
  gamm_hc_lft, gamm_hc_gamm,
  hc_age_lme, mem_age_lme
)

# Define a function to extract p-values
extract_p_value <- function(model) {
  if ("p_value" %in% names(model)) {
    return(model[["p_value"]])
  } else if ("p_value_gamm" %in% names(model)) {
    return(model[["p_value_gamm"]])
  } else if ("p_value_lrt" %in% names(model)) {
    return(model[["p_value_lrt"]])
  } else {
    return(NULL)
  }
}

adjust_p_values <- function(models, model_names) {
  p_values <- unlist(lapply(models, extract_p_value))
  p_adjusted <- p.adjust(p_values, method = "fdr")
  data.frame(
    model_name = model_names,
    original_p_value = format(p_values, digits = 4, scientific = FALSE, na.encode = TRUE),
    adjusted_p_value = format(p_adjusted, digits = 4, scientific = FALSE, na.encode = TRUE)
  )
}

# Adjust p-values for each group
cohort_comp_p_values <- adjust_p_values(
  cohort_comp, 
  c("age_model", "age_model_fu", "chisq_test", "edu_model", "mmse_model", "fu_time_model", "fu_obs_model")
)
memory_hippocampus_p_values <- adjust_p_values(
  memory_hippocampus_edu_models,
  c("emhp_merged", "model_baseline", "model_followup")
)
education_cr_p_values <- adjust_p_values(
  education_cr_models,
  c("edu_cr", "edu_cr_int", "edu_mem", "edu_cr_hc")
)
cr_sensitivity_p_values <- adjust_p_values(
  cr_sensitivity_models,
  c("crbm_corr")
)
fc_long_cr_ecn_p_values <- adjust_p_values(
  fc_long_cr_ecn,
  c("cr_ecn_stats", "cr_ec", "cr_ec_mem")
)
fc_long_cr_dmn_p_values <- adjust_p_values(
  fc_long_cr_dmn,
  c("cr_dmn_stats", "cr_dmn", "cr_dmn_mem")
)
fc_long_cr_sn_p_values <- adjust_p_values(
  fc_long_cr_sn,
  c("cr_sn_stats", "cr_sn", "cr_sn_mem")
)
model_comp_general_p_values <- adjust_p_values(
  model_comp_general_models,
  c("gamm_memory_lft", "gamm_memory_gamm", "gamm_hc_lft", "gamm_hc_gamm", "hc_age_lme", "mem_age_lme")
)

# Combine all results
p_value_df <- rbind(
  cohort_comp_p_values,
  memory_hippocampus_p_values,
  education_cr_p_values,
  cr_sensitivity_p_values,
  fc_long_cr_ecn_p_values, 
  fc_long_cr_dmn_p_values, 
  fc_long_cr_sn_p_values, 
  model_comp_general_p_values
)

# Output the combined p-values
write.csv(p_value_df, file = "/tsd/p274/data/durable/projects/p027-cr_bm/p_value.csv")
```
